fix: Corrected the parameter in findByUrlId method
In the MusicRepository class, the findByUrlId method was previously searching for Music entities based on the 'url' field. This was causing errors as the method should have been searching based on the 'urlId' field. This commit corrects the parameter in the findByUrlId method from 'url' to 'urlId', fixing the issue.